refactor: split large files into focused modules and improve test infrastructure
Extract code from large files into dedicated modules (TokenStorageError, FieldValue+Codable, FilterBuilder string/list filters, CloudKitResponseProcessor+Changes, CloudKitService+AssetUpload) and restructure test helpers into separate files (ResponseConfig, ResponseProvider, ValidationErrorType, MockRecordManagingService, TestRecord, test extension files).
