+++ Sources/MistKit/Authentication/TokenStorageError.swift
@@ -1,5 +1,5 @@
 //
-//  ZoneInfo.swift
+//  TokenStorageError.swift
 //  MistKit
 //
 //  Created by Leo Dion.
@@ -27,21 +27,37 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Zone information from CloudKit
-public struct ZoneInfo: Encodable, Sendable {
-  /// The zone name
-  public let zoneName: String
-  /// The owner record name
-  public let ownerRecordName: String?
-  /// The zone capabilities.
-  /// Note: always empty — CloudKit Web Services zone responses do not include
-  /// capabilities in the current OpenAPI schema.
-  public let capabilities: [String]
+import Foundation
 
-  /// Initialize zone information
-  public init(zoneName: String, ownerRecordName: String?, capabilities: [String]) {
-    self.zoneName = zoneName
-    self.ownerRecordName = ownerRecordName
-    self.capabilities = capabilities
+/// Errors that can occur during token storage operations
+public enum TokenStorageError: Error, LocalizedError, Sendable {
+  /// Storage operation failed
+  case storageFailed(reason: String)
+
+  /// Credentials not found
+  case notFound(identifier: String?)
+
+  /// Access denied to storage
+  case accessDenied
+
+  /// Storage corrupted or invalid format
+  case corruptedStorage
+
+  /// A localized message describing what error occurred
+  public var errorDescription: String? {
+    switch self {
+    case .storageFailed(let reason):
+      return "Token storage failed: \(reason)"
+    case .notFound(let identifier):
+      if let identifier = identifier {
+        return "Credentials not found for identifier: \(identifier)"
+      } else {
+        return "No credentials found"
+      }
+    case .accessDenied:
+      return "Access denied to token storage"
+    case .corruptedStorage:
+      return "Token storage is corrupted or in invalid format"
+    }
   }
 }
