+++ Sources/MistKit/Extensions/OpenAPI/Components.Schemas.ListValuePayload+MistKit.swift
@@ -32,59 +32,88 @@ internal import Foundation
 extension Components.Schemas.ListValuePayload {
   /// Initialize from MistKit FieldValue for list elements
   internal init(from fieldValue: FieldValue) {
-    switch fieldValue {
-    case .string(let value):
-      self = .StringValue(value)
-    case .int64(let value):
-      self = .Int64Value(Int64(value))
-    case .double(let value):
-      self = .DoubleValue(value)
-    case .bytes(let value):
-      self = .BytesValue(value)
-    case .date(let value):
-      let milliseconds = value.timeIntervalSince1970 * 1_000
-      self = .DateValue(milliseconds)
-    case .location(let location):
-      let locationValue = Components.Schemas.LocationValue(
-        latitude: location.latitude,
-        longitude: location.longitude,
-        horizontalAccuracy: location.horizontalAccuracy,
-        verticalAccuracy: location.verticalAccuracy,
-        altitude: location.altitude,
-        speed: location.speed,
-        course: location.course,
-        timestamp: location.timestamp.map { $0.timeIntervalSince1970 * 1_000 }
-      )
-      self = .LocationValue(locationValue)
-    case .reference(let reference):
-      let action: Components.Schemas.ReferenceValue.actionPayload?
-      switch reference.action {
-      case .some(.deleteSelf):
-        action = .DELETE_SELF
-      case .some(.none):
-        action = .NONE
-      case nil:
-        action = nil
-      }
-      let referenceValue = Components.Schemas.ReferenceValue(
-        recordName: reference.recordName,
-        action: action
-      )
-      self = .ReferenceValue(referenceValue)
-    case .asset(let asset):
-      let assetValue = Components.Schemas.AssetValue(
-        fileChecksum: asset.fileChecksum,
-        size: asset.size,
-        referenceChecksum: asset.referenceChecksum,
-        wrappingKey: asset.wrappingKey,
-        receipt: asset.receipt,
-        downloadURL: asset.downloadURL
-      )
-      self = .AssetValue(assetValue)
-    case .list(let nestedList):
-      // Recursively convert nested lists
-      let nestedPayloads = nestedList.map { Self(from: $0) }
-      self = .ListValue(nestedPayloads)
+    if let payload = Self.makeScalarPayload(from: fieldValue) {
+      self = payload
+    } else {
+      self = Self.makeComplexPayload(from: fieldValue)
     }
+  }
+
+  private static func makeScalarPayload(from fieldValue: FieldValue) -> Self? {
+    if case .string(let value) = fieldValue {
+      return .StringValue(value)
+    }
+    if case .int64(let value) = fieldValue {
+      return .Int64Value(Int64(value))
+    }
+    if case .double(let value) = fieldValue {
+      return .DoubleValue(value)
+    }
+    if case .bytes(let value) = fieldValue {
+      return .BytesValue(value)
+    }
+    if case .date(let value) = fieldValue {
+      return .DateValue(value.timeIntervalSince1970 * 1_000)
+    }
+    return nil
+  }
+
+  private static func makeComplexPayload(from fieldValue: FieldValue) -> Self {
+    switch fieldValue {
+    case .location(let location):
+      return .LocationValue(makeLocationValue(location))
+    case .reference(let reference):
+      return .ReferenceValue(makeReferenceValue(reference))
+    case .asset(let asset):
+      return .AssetValue(makeAssetValue(asset))
+    case .list(let nestedList):
+      return .ListValue(nestedList.map { Self(from: $0) })
+    default:
+      return .ListValue([])
+    }
+  }
+
+  private static func makeLocationValue(_ location: FieldValue.Location)
+    -> Components.Schemas.LocationValue
+  {
+    Components.Schemas.LocationValue(
+      latitude: location.latitude,
+      longitude: location.longitude,
+      horizontalAccuracy: location.horizontalAccuracy,
+      verticalAccuracy: location.verticalAccuracy,
+      altitude: location.altitude,
+      speed: location.speed,
+      course: location.course,
+      timestamp: location.timestamp.map { $0.timeIntervalSince1970 * 1_000 }
+    )
+  }
+
+  private static func makeReferenceValue(_ reference: FieldValue.Reference)
+    -> Components.Schemas.ReferenceValue
+  {
+    let action: Components.Schemas.ReferenceValue.actionPayload?
+    switch reference.action {
+    case .some(.deleteSelf):
+      action = .DELETE_SELF
+    case .some(.none):
+      action = .NONE
+    case nil:
+      action = nil
+    }
+    return Components.Schemas.ReferenceValue(
+      recordName: reference.recordName,
+      action: action
+    )
+  }
+
+  private static func makeAssetValue(_ asset: FieldValue.Asset) -> Components.Schemas.AssetValue {
+    Components.Schemas.AssetValue(
+      fileChecksum: asset.fileChecksum,
+      size: asset.size,
+      referenceChecksum: asset.referenceChecksum,
+      wrappingKey: asset.wrappingKey,
+      receipt: asset.receipt,
+      downloadURL: asset.downloadURL
+    )
   }
 }
