+++ Sources/MistKit/FieldValue.swift
@@ -134,112 +134,4 @@ public enum FieldValue: Codable, Equatable, Sendable {
       self.downloadURL = downloadURL
     }
   }
-
-  /// Conversion factor from seconds to milliseconds for CloudKit timestamps
-  private static let millisecondsPerSecond: Double = 1_000
-
-  // MARK: - Codable
-  /// Initialize field value from decoder
-  public init(from decoder: any Decoder) throws {
-    let container = try decoder.singleValueContainer()
-
-    if let value = try Self.decodeBasicTypes(from: container) {
-      self = value
-      return
-    }
-
-    if let value = try Self.decodeComplexTypes(from: container) {
-      self = value
-      return
-    }
-
-    throw DecodingError.dataCorruptedError(
-      in: container,
-      debugDescription: "Unable to decode FieldValue"
-    )
-  }
-
-  /// Decode basic field value types (string, int64, double)
-  private static func decodeBasicTypes(from container: any SingleValueDecodingContainer) throws
-    -> FieldValue?
-  {
-    if let value = try? container.decode(String.self) {
-      return .string(value)
-    }
-    if let value = try? container.decode(Int.self) {
-      return .int64(value)
-    }
-    if let value = try? container.decode(Double.self) {
-      return .double(value)
-    }
-    return nil
-  }
-
-  /// Decode complex field value types (list, location, reference, asset, date)
-  private static func decodeComplexTypes(from container: any SingleValueDecodingContainer) throws
-    -> FieldValue?
-  {
-    if let value = try? container.decode([FieldValue].self) {
-      return .list(value)
-    }
-    if let value = try? container.decode(Location.self) {
-      return .location(value)
-    }
-    if let value = try? container.decode(Reference.self) {
-      return .reference(value)
-    }
-    if let value = try? container.decode(Asset.self) {
-      return .asset(value)
-    }
-    // Try to decode as date (milliseconds since epoch)
-    if let value = try? container.decode(Double.self) {
-      return .date(Date(timeIntervalSince1970: value / Self.millisecondsPerSecond))
-    }
-    return nil
-  }
-
-  /// Encode field value to encoder
-  public func encode(to encoder: any Encoder) throws {
-    var container = encoder.singleValueContainer()
-    try encodeValue(to: &container)
-  }
-
-  // swiftlint:disable:next cyclomatic_complexity
-  private func encodeValue(to container: inout any SingleValueEncodingContainer) throws {
-    switch self {
-    case .string(let val), .bytes(let val):
-      try container.encode(val)
-    case .int64(let val):
-      try container.encode(val)
-    case .double(let val):
-      try container.encode(val)
-    case .date(let val):
-      try container.encode(val.timeIntervalSince1970 * Self.millisecondsPerSecond)
-    case .location(let val):
-      try container.encode(val)
-    case .reference(let val):
-      try container.encode(val)
-    case .asset(let val):
-      try container.encode(val)
-    case .list(let val):
-      try container.encode(val)
-    }
-  }
-}
-
-// MARK: - Helper Methods
-
-extension FieldValue {
-  /// Create an int64 FieldValue from a Bool
-  ///
-  /// CloudKit represents booleans as INT64 (0/1) on the wire.
-  /// Creates a FieldValue from a Swift Bool value.
-  ///
-  /// This initializer converts Swift Bool to the appropriate int64 representation
-  /// used by CloudKit (1 for true, 0 for false).
-  ///
-  /// - Parameter booleanValue: The boolean value to convert
-  public init(booleanValue: Bool) {
-    self = .int64(booleanValue ? 1 : 0)
-  }
 }
