+++ Sources/MistKit/Service/CloudKitResponseProcessor.swift
@@ -199,94 +199,4 @@ internal struct CloudKitResponseProcessor {
       throw CloudKitError.invalidResponse
     }
   }
-
-  /// Process fetchRecordChanges response
-  /// - Parameter response: The response to process
-  /// - Returns: The extracted changes response data
-  /// - Throws: CloudKitError for various error conditions
-  internal func processFetchRecordChangesResponse(_ response: Operations.fetchRecordChanges.Output)
-    async throws(CloudKitError) -> Components.Schemas.ChangesResponse
-  {
-    // Check for errors first
-    if let error = CloudKitError(response) {
-      throw error
-    }
-
-    // Must be .ok case - extract data
-    switch response {
-    case .ok(let okResponse):
-      switch okResponse.body {
-      case .json(let changesData):
-        return changesData
-      }
-    default:
-      // Should never reach here since all errors are handled above
-      assertionFailure("Unexpected response case after error handling")
-      throw CloudKitError.invalidResponse
-    }
-  }
-
-  /// Process discoverUserIdentities response
-  internal func processDiscoverUserIdentitiesResponse(
-    _ response: Operations.discoverUserIdentities.Output
-  ) async throws(CloudKitError) -> Components.Schemas.DiscoverResponse {
-    if let error = CloudKitError(response) {
-      throw error
-    }
-    switch response {
-    case .ok(let okResponse):
-      switch okResponse.body {
-      case .json(let discoverData):
-        return discoverData
-      }
-    default:
-      assertionFailure("Unexpected response case after error handling")
-      throw CloudKitError.invalidResponse
-    }
-  }
-
-  /// Process uploadAssets response
-  /// - Parameter response: The response to process
-  /// - Returns: The extracted asset upload response data
-  /// - Throws: CloudKitError for various error conditions
-  internal func processUploadAssetsResponse(_ response: Operations.uploadAssets.Output)
-    async throws(CloudKitError) -> Components.Schemas.AssetUploadResponse
-  {
-    // Check for errors first
-    if let error = CloudKitError(response) {
-      throw error
-    }
-
-    // Must be .ok case - extract data
-    switch response {
-    case .ok(let okResponse):
-      switch okResponse.body {
-      case .json(let uploadData):
-        return uploadData
-      }
-    default:
-      // Should never reach here since all errors are handled above
-      assertionFailure("Unexpected response case after error handling")
-      throw CloudKitError.invalidResponse
-    }
-  }
-
-  /// Process fetchZoneChanges response
-  internal func processFetchZoneChangesResponse(_ response: Operations.fetchZoneChanges.Output)
-    async throws(CloudKitError) -> Components.Schemas.ZoneChangesResponse
-  {
-    if let error = CloudKitError(response) {
-      throw error
-    }
-    switch response {
-    case .ok(let okResponse):
-      switch okResponse.body {
-      case .json(let changesData):
-        return changesData
-      }
-    default:
-      assertionFailure("Unexpected response case after error handling")
-      throw CloudKitError.invalidResponse
-    }
-  }
 }
