+++ Tests/MistKitTests/Client/MistKitClientTests.swift
@@ -192,50 +192,4 @@ internal struct MistKitClientTests {
       }
     }
   }
-
-  @Test("MistKitClient rejects ServerToServerAuthManager with shared database")
-  internal func serverToServerWithSharedDatabase() throws {
-    guard #available(macOS 11.0, iOS 14.0, tvOS 14.0, watchOS 7.0, *) else {
-      return
-    }
-
-    let privateKeyPEM = """
-      -----BEGIN PRIVATE KEY-----
-      MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgevZzL1gdAFr88hb2
-      OF/2NxApJCzGCEDdfSp6VQO30hyhRANCAAQRWz+jn65BtOMvdyHKcvjBeBSDZH2r
-      1RTwjmYSi9R/zpBnuQ4EiMnCqfMPWiZqB4QdbAd0E7oH50VpuZ1P087G
-      -----END PRIVATE KEY-----
-      """
-
-    let tokenManager = try ServerToServerAuthManager(
-      keyID: String(repeating: "0", count: 64),
-      pemString: privateKeyPEM
-    )
-
-    let config = MistKitConfiguration(
-      container: "iCloud.com.example.app",
-      environment: .development,
-      database: .shared,
-      apiToken: ""
-    )
-
-    let transport = MockTransport()
-
-    do {
-      _ = try MistKitClient(
-        configuration: config,
-        tokenManager: tokenManager,
-        transport: transport
-      )
-      Issue.record("Expected TokenManagerError for server-to-server with shared database")
-    } catch let error as TokenManagerError {
-      if case .invalidCredentials = error {
-        // Success
-      } else {
-        Issue.record("Expected invalidCredentials error, got \(error)")
-      }
-    }
-  }
-
 }
-
