+++ Tests/MistKitTests/Protocols/RecordManagingTests.swift
@@ -27,44 +27,9 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import Foundation
 import Testing
 
 @testable import MistKit
-
-/// Mock implementation of RecordManaging for testing
-internal actor MockRecordManagingService: RecordManaging {
-  internal var queryCallCount = 0
-  internal var executeCallCount = 0
-  internal var lastExecutedOperations: [RecordOperation] = []
-  internal var batchSizes: [Int] = []
-  internal var recordsToReturn: [RecordInfo] = []
-
-  internal func queryRecords(recordType: String) async throws -> [RecordInfo] {
-    queryCallCount += 1
-    return recordsToReturn
-  }
-
-  internal func executeBatchOperations(_ operations: [RecordOperation], recordType: String)
-    async throws
-  {
-    executeCallCount += 1
-    batchSizes.append(operations.count)
-    lastExecutedOperations.append(contentsOf: operations)
-  }
-
-  internal func reset() {
-    queryCallCount = 0
-    executeCallCount = 0
-    lastExecutedOperations = []
-    batchSizes = []
-    recordsToReturn = []
-  }
-
-  internal func setRecordsToReturn(_ records: [RecordInfo]) {
-    recordsToReturn = records
-  }
-}
 
 @Suite("RecordManaging Protocol")
 internal enum RecordManagingTests {}
