+++ Tests/MistKitTests/Service/CloudKitDiscoverUserIdentitiesTests.swift
@@ -1,5 +1,5 @@
 //
-//  ZoneInfo.swift
+//  CloudKitDiscoverUserIdentitiesTests.swift
 //  MistKit
 //
 //  Created by Leo Dion.
@@ -7,7 +7,7 @@
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the “Software”), to deal in the Software without
+//  files (the "Software"), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -27,21 +27,11 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Zone information from CloudKit
-public struct ZoneInfo: Encodable, Sendable {
-  /// The zone name
-  public let zoneName: String
-  /// The owner record name
-  public let ownerRecordName: String?
-  /// The zone capabilities.
-  /// Note: always empty — CloudKit Web Services zone responses do not include
-  /// capabilities in the current OpenAPI schema.
-  public let capabilities: [String]
+import Foundation
+import Testing
 
-  /// Initialize zone information
-  public init(zoneName: String, ownerRecordName: String?, capabilities: [String]) {
-    self.zoneName = zoneName
-    self.ownerRecordName = ownerRecordName
-    self.capabilities = capabilities
-  }
-}
+@testable import MistKit
+
+@Suite(
+  "CloudKitService DiscoverUserIdentities Operations", .enabled(if: Platform.isCryptoAvailable))
+internal enum CloudKitDiscoverUserIdentitiesTests {}
